Clean up LeaderboardService and types
