--- common/src/main/java/com/wynntils/models/leaderboard/type/LeaderboardBadge.java
@@ -1,10 +1,0 @@
-/*
- * Copyright © Wynntils 2023.
- * This file is released under LGPLv3. See LICENSE for full license details.
- */
-package com.wynntils.models.leaderboard.type;
-
-public record LeaderboardBadge(int uOffset, int vOffset) {
-    public static final int WIDTH = 19;
-    public static final int HEIGHT = 17;
-}
